functions/pubsub_to_cassandra/src/requirements.txt
functions-framework==3.*
cassandra-driver==3.29.0
google-cloud-logging==3.*